latest changes new

+++ src/endpoints/tasks.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, Request, status,HTTPException,Depends
 from src.endpoints.auth_beare import JWTBearer
 from typing import List
-from src.model.tasks import Task, UpdateTask,CreateTask
+from src.model.tasks import Task, UpdateTask,AllTask
 from fastapi.encoders import jsonable_encoder
 from bson import ObjectId
 import uuid
@@ -18,8 +18,8 @@ def get_collection_task(request: Request):
 router = APIRouter(prefix="/Tasks",
     tags=["Tasks"])
 
-@router.post("/", response_description="Create Task", dependencies=dependency_JWT, status_code=status.HTTP_201_CREATED, response_model=CreateTask)
-def create_task(request: Request, task: CreateTask):
+@router.post("/", response_description="Create Task", dependencies=dependency_JWT, status_code=status.HTTP_201_CREATED, response_model=Task)
+def create_task(request: Request, task: Task):
     tasks = jsonable_encoder(task)
     global current_user
     current_user=get_current_user_id()
@@ -29,12 +29,15 @@ def create_task(request: Request, task: CreateTask):
     return created_task
 
 
-@router.put("/{id}/", response_description="Update Task", dependencies=dependency_JWT, response_model=Task)
+@router.put("/{id}/", response_description="Update Task", dependencies=dependency_JWT, response_model=UpdateTask)
 def update_task(request: Request, id: str, task: UpdateTask):
     tasks = {k: v for k, v in task.dict().items() if v is not None}
     print(tasks)
+    global current_user
+    current_user=get_current_user_id()
+    print(current_user)
     if len(tasks) >= 1:
-        update_result = get_collection_task(request).update_one({"_id": ObjectId(id)}, {"$set": tasks})
+        update_result = get_collection_task(request).update_one({"_id": ObjectId(id),"user_id":current_user}, {"$set": tasks})
         if update_result.modified_count == 0:
             raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Task not found!")
     if (existing_book := get_collection_task(request).find_one({"_id": ObjectId(id)})) is not None:
@@ -44,7 +47,7 @@ def update_task(request: Request, id: str, task: UpdateTask):
 
 
 
-@router.get("/", response_description="All Tasks", response_model=List[Task])
+@router.get("/", response_description="All Tasks",dependencies=dependency_JWT, response_model=List[Task])
 def list_task(request: Request,limit:int):
     global current_user
     current_user=get_current_user_id()
@@ -53,16 +56,19 @@ def list_task(request: Request,limit:int):
     return task
 
 
-@router.get("/{id}/", response_description="Fetch Task By ID", response_model=Task)
+@router.get("/{id}/", response_description="Fetch Task By ID", dependencies=dependency_JWT, response_model=Task)
 def list_task_by_id(request: Request, id: str):
-    if (task := get_collection_task(request).find_one({"_id": ObjectId(id)})):
+    if (task := get_collection_task(request).find_one({"_id": ObjectId(id),"user_id":current_user})):
         return task
     raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Task not found!")
 
 
 @router.delete("/{id}",  dependencies=dependency_JWT, response_description="Delete Task")
 def delete_task(request: Request, id: str):
-    deleted_task = get_collection_task(request).delete_one({"_id": ObjectId(id)})
+    global current_user
+    current_user=get_current_user_id()
+    print(current_user)
+    deleted_task = get_collection_task(request).delete_one({"_id": ObjectId(id),"user_id":current_user})
     if deleted_task.deleted_count == 1:
         return f"Task with ID {id} deleted sucessfully!"
     raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Task not found!")
